Use newly created function to create new users
The password for the writer user will be stored in a non-writeable file
that's only readable by the user running the initiation script in the
credentials directory, and the password for the admin account will be
stored in a file that's neither read nor writeable in the same
directory. The reason for this, is that the admin account is not
supposed to being used after setup, but it might be nice to have for
debugging purposes, which is why the password is stored in a file that's
essentially only readable if you have root.