Prevent from looping on scraping articles
Prior the program could get stuck in a loop if one article caused an
error when scraping it. The reason for this, is that the OG tags and
meta information for the article would be scraped and stored, but if the
program crashes when scraping the article for creating the markdown file
the article wouldn't be marked as scraped, and then when starting the
program again, it would start trying to scrape that article and crash
again. This has been fixed by wrapping the function for scraping the
articles marked as not scraped in a try except, and notifying the user
when encountering an error.